+++ app/Http/Controllers/EditUsersController.php
@@ -2,7 +2,7 @@
 
 namespace App\Http\Controllers;
 
-class UserController extends Controller
+class EditUsersController extends Controller
 {
     public function __construct() {
         $this->middleware('auth');
@@ -11,6 +11,6 @@ class UserController extends Controller
 
     public function index()
     {
-        return view('user');
+        return view('edit-users');
     }
 }