+++ resources/lang/en/messages.php
@@ -58,5 +58,8 @@ return [
         'add_item' => 'Add item',
         'deletion_title' => 'Confirm deletion',
         'deletion_warning' => "Deleting an item should only be done if the item doesn't, and never did, correspond to an actual inventory item.\nIf the item was lost, you should rather change its status.\nAfter deleting an item, all its borrowings will be deleted as well from the database.\nThis deletion can't be undone."
+    ],
+    'edit_users' => [
+        'change_role' => 'Change role'
     ]
 ];