+++ changelog.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 * Edit inventory view
     * Minimal width of tables for better responsiveness.
+* Edit users view
+    * Layout.
     
 ### Removed