+++ changelog.md
@@ -5,7 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [0.11.0] - Unreleased
+## [0.12.0] - Unreleased
+
+### Added
+
+* Register validation
+    * Promotion must now be within an acceptable range.
+
+## [0.11.0] - 2019-10-01
 
 ### Added